+++ src/apps/coreutils/wc/loop.cc
@@ -8,8 +8,8 @@ static inline int isspace(int c) {
     return isblank(c) || c == '\v' || c == '\f' || c == '\r' || c == '\n';
 }
 
-void count(const char *buffer, long res, long *lines, long *words, int *last_space) {
-    for(long i = 0; i < res; ++i) {
+void count(const char *buffer, size_t res, long *lines, long *words, int *last_space) {
+    for(size_t i = 0; i < res; ++i) {
         if(buffer[i] == '\n')
             (*lines)++;
         int space = isspace(buffer[i]);
